+++ src/main/kotlin/io/snyk/plugin/analytics/AnalyticsSender.kt
@@ -1,6 +1,7 @@
 package io.snyk.plugin.analytics
 
 import com.intellij.openapi.Disposable
+import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
 import io.snyk.plugin.ui.toolwindow.SnykPluginDisposable
 import org.jetbrains.concurrency.runAsync
@@ -9,20 +10,20 @@ import snyk.common.lsp.analytics.AbstractAnalyticsEvent
 import java.util.LinkedList
 import java.util.concurrent.ConcurrentLinkedQueue
 
-class AnalyticsSender : Disposable {
+class AnalyticsSender(val project: Project) : Disposable {
     private var disposed: Boolean = false
 
     // left = event, right = callback function
     private val eventQueue = ConcurrentLinkedQueue<Pair<AbstractAnalyticsEvent, () -> Unit>>()
 
     init {
-        Disposer.register(SnykPluginDisposable.getInstance(), this)
+        Disposer.register(SnykPluginDisposable.getInstance(project), this)
         start()
     }
 
     private fun start() {
         runAsync {
-            val lsw = LanguageServerWrapper.getInstance()
+            val lsw = LanguageServerWrapper.getInstance(project)
             while (!disposed) {
                 if (eventQueue.isEmpty() || lsw.notAuthenticated()) {
                     Thread.sleep(1000)
@@ -49,9 +50,9 @@ class AnalyticsSender : Disposable {
         private var instance: AnalyticsSender? = null
 
         @JvmStatic
-        fun getInstance(): AnalyticsSender {
+        fun getInstance(project: Project): AnalyticsSender {
             if (instance == null) {
-                instance = AnalyticsSender()
+                instance = AnalyticsSender(project)
             }
             return instance as AnalyticsSender
         }