+++ src/main/kotlin/snyk/common/AnnotatorCommon.kt
@@ -1,6 +1,7 @@
 package snyk.common
 
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
@@ -10,15 +11,15 @@ import io.snyk.plugin.events.SnykSettingsListener
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.refreshAnnotationsForOpenFiles
 import snyk.common.lsp.LanguageServerWrapper
-
-object AnnotatorCommon {
+@Service(Service.Level.PROJECT)
+class AnnotatorCommon(val project: Project) {
     val logger = logger<AnnotatorCommon>()
 
     fun prepareAnnotate(psiFile: PsiFile?) {
         logger.debug("Preparing annotation for $psiFile")
 
         // trigger LS initialization if not already done, we consciously don't check the result here
-        LanguageServerWrapper.getInstance().ensureLanguageServerInitialized()
+        LanguageServerWrapper.getInstance(project).ensureLanguageServerInitialized()
 
         // todo: review later if any way to provide up-to-date context for CLI scans is available
         // force saving here will break some user's workflow: https://github.com/snyk/snyk-intellij-plugin/issues/324
@@ -27,7 +28,7 @@ object AnnotatorCommon {
     fun isSeverityToShow(severity: Severity): Boolean =
         pluginSettings().hasSeverityEnabled(severity) || severity == Severity.UNKNOWN
 
-    fun initRefreshing(project: Project) {
+    fun initRefreshing() {
         if (project.isDisposed) return
         if (ApplicationManager.getApplication().isDisposed) return
         logger.debug("Initializing annotations refresh listener")